pyproject.toml
[project]
name = "nanochat"
version = "0.2.0"
description = "the minimal full-stack ChatGPT clone — JAX/TPU edition"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.0.0",
    "fastapi>=0.117.1",
    "flax>=0.10.0",
    "jax[tpu]>=0.6.0",
    "optax>=0.2.4",
    "orbax-checkpoint>=0.11.0",
    "psutil>=7.1.0",
    "rustbpe>=0.1.0",
    "tiktoken>=0.11.0",
    "tokenizers>=0.22.0",
    "uvicorn>=0.36.0",
    "wandb>=0.21.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "pytest>=8.0.0",
    "python-dotenv>=1.2.1",
    "transformers>=4.57.3",
]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.uv]
default-groups = []

# jax[tpu] requires the libtpu wheel from Google's index
[tool.uv.sources]
jax = [
    { index = "google-jax-tpu" },
]

[[tool.uv.index]]
name = "google-jax-tpu"
url = "https://storage.googleapis.com/jax-releases/libtpu_releases.html"
explicit = true
